Cleanup import violation of cmk_base in cmk.gui code
